checkout model and routes

+++ routes/order.js
@@ -5,8 +5,8 @@ const Order = require("../models/order");
 const Cart = require("../models/cart");
 const User = require("../models/user");
 
-router.get("/", async (req, res) => {
-  const owner = req.body._id;
+router.get("/:id", async (req, res) => {
+  const owner = req.params.id;
   console.log(owner);
   try {
     const order = await Order.find({ owner: owner }).sort({ date: -1 });
@@ -19,22 +19,34 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.post("/checkout", async (req, res) => {
+router.post("/checkout/:id", async (req, res) => {
   try {
-    const owner = req.body._id;
+    const owner = req.params.id;
+    const name = req.body.name;
+    const mail = req.body.mail;
+    const address = req.body.address;
+    const city = req.body.city;
+    const zip = req.body.zip;
+
     //find cart and user
     let cart = await Cart.findOne({ owner });
-    if (cart) {
+
+    if (!cart) {
+      res.status(400).send("Cart is empty");
+    }
+
+    if (cart && cart.products !== []) {
       await Cart.findByIdAndDelete({ _id: cart.id });
       const newOrder = await Order.create({
         owner,
+        credentials: [{ name, mail, address, city, zip }],
         products: cart.products,
         bill: cart.bill,
         status: "pending",
       });
       return res.status(201).send(newOrder);
     } else {
-      res.status(400).send("No cart found");
+      res.status(400).send("Something wrong, please try again.");
     }
   } catch (error) {
     console.log(error);
